Quitting when ending #3
To solve this problem I decided to just not join on the twitch wrapper thread. The join waits until that thread is done, however that thread will never be done because it runs an infinite loop. Instead of setting up some kind of commands to tell the twitch wrapper to quit, I'm just not joining on that thread anymore. Now when the game quits the main function ends which immediately ends all threads spawned by the application.

I'm not completely sure that there won't be any negative side affects, but it seems to work well with the tests that I've done.